requirements.txt
pandas
numpy
scikit-learn
google-cloud-bigquery
fastapi
uvicorn
streamlit
requests